requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
alembic
python-jose
passlib[bcrypt]
python-multipart
pydantic-settings
email-validator
reportlab